Fix Radsec Regression by using Ubuntu instead of Alpine (#46)
A bug in Openssl and the latest version of Alpine prevents Radsec from
working. More on this here: http://lists.freeradius.org/pipermail/freeradius-users/2021-September/100708.html

Downgrading Alpine is not an option as we lose the Python3 module
integration required by the policy engine.

Ubuntu 20.4 supports both Radsec and the Policy Engine.

Tests temporarily disabled as they are being ported to the integration
test repo.