Run FreeRadius as non root user
This improves the security posture of the servers, removing the ability
to run sudo or access restricted areas of the filesystem.

NGINX removed for now, this was used for local OCSP tests, which is
disabled. OCSP automated testing will be refactored to pull the endpoint
out into a new container.