Install tshark for packet capture
This conditionally installs wireshark for packet capturing if not local
development environment.

A future pull request will narrow down this condition even more to only
look for the debug flag from SSM Parameter store.

Move aws cli installation to a script with a conditional to not install
it locally. It is slow to install and unused during automated testing or
local development.