Ability to force an Access-Reject from Policy Engine
By returning RLM_MODULE_FAIL instead of RLM_MODULE_OK, we are able to
short curcuit the authentication at the post-auth stage.

This looks for a response of "Post-Auth-Type == Reject", which will send
back an access reject even though the authenticate stage resulted in
Access-Accept. This is a better user experience than sending devices to
a blackhole VLAN.